moved components into structure, again, consistency
